requirements.txt
numpy
pandas
matplotlib
seaborn
scipy
streamlit
streamlit_option_menu
streamlit_lottie
python-dotenv
google-generativeai
langchain
langchain-openai
langchain-core
langchain-community
langchain-experimental
langchain-huggingface
langchain-text-splitters